[6799] Move palette DTO in a dedicated package
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6799
