+++ packages/diagrams/backend/sirius-components-collaborative-diagrams/pom.xml
@@ -44,6 +44,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.eclipse.sirius</groupId>
+			<artifactId>sirius-components-palette</artifactId>
+			<version>2026.7.2</version>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.sirius</groupId>
 			<artifactId>sirius-components-collaborative</artifactId>
 			<version>2026.7.2</version>
 		</dependency>
