+++ packages/diagrams/backend/sirius-components-collaborative-diagrams/src/main/resources/schema/diagram.graphqls
@@ -388,34 +388,7 @@ type Size {
   width: Float!
 }
 
-type Palette {
-  id: ID! @deprecated(reason: "https://github.com/eclipse-sirius/sirius-web/issues/6470")
-  quickAccessTools: [Tool]!
-  paletteEntries: [PaletteEntry]!
-}
-
-union PaletteEntry =
-  ToolSection
-  | SingleClickOnDiagramElementTool
-  | SingleClickOnTwoDiagramElementsTool
-  | PaletteDivider
-
-type PaletteDivider {
-  id: ID!
-}
-
-type ToolSection {
-  id: ID!
-  label: String!
-  iconURL: [String!]!
-  tools: [Tool]!
-}
-
-interface Tool {
-  id: ID!
-  label: String!
-  iconURL: [String!]!
-}
+extend union PaletteEntry = SingleClickOnDiagramElementTool | SingleClickOnTwoDiagramElementsTool
 
 type SingleClickOnDiagramElementTool implements Tool {
   id: ID!
