+++ packages/palette/backend/sirius-components-palette/src/main/java/org/eclipse/sirius/components/palette/dto/ITool.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Obeo.
+ * Copyright (c) 2023, 2026 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,18 +10,18 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.components.collaborative.diagrams.dto;
+package org.eclipse.sirius.components.palette.dto;
 
-import java.util.Objects;
+import java.util.List;
 
 /**
- * Represents a divider between tools or tool sections in the palette.
+ * Interface implemented by all tools.
  *
- * @author fbarbin
+ * @author mcharfadi
  */
-public record PaletteDivider(String id) implements IPaletteEntry {
+public interface ITool extends IPaletteEntry {
 
-    public PaletteDivider {
-        Objects.requireNonNull(id);
-    }
+    String label();
+
+    List<String> iconURL();
 }
