+++ packages/palette/backend/sirius-components-palette/src/main/java/org/eclipse/sirius/components/palette/dto/Palette.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.components.collaborative.diagrams.dto;
+package org.eclipse.sirius.components.palette.dto;
 
 import java.util.List;
 import java.util.Objects;
