+++ packages/view/backend/sirius-components-view-emf/src/main/java/org/eclipse/sirius/components/view/emf/diagram/api/IGroupPaletteToolsProvider.java
@@ -15,8 +15,8 @@ package org.eclipse.sirius.components.view.emf.diagram.api;
 import java.util.List;
 
 import org.eclipse.sirius.components.collaborative.diagrams.DiagramContext;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.ITool;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.ToolSection;
+import org.eclipse.sirius.components.palette.dto.ITool;
+import org.eclipse.sirius.components.palette.dto.ToolSection;
 import org.eclipse.sirius.components.diagrams.description.IDiagramElementDescription;
 
 /**
