+++ packages/view/backend/sirius-components-view-emf/src/main/java/org/eclipse/sirius/components/view/emf/diagram/tools/DiagramPaletteProvider.java
@@ -19,10 +19,10 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.eclipse.emf.ecore.util.EcoreUtil;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.IPaletteEntry;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.ITool;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.Palette;
-import org.eclipse.sirius.components.collaborative.diagrams.dto.ToolSection;
+import org.eclipse.sirius.components.palette.dto.IPaletteEntry;
+import org.eclipse.sirius.components.palette.dto.ITool;
+import org.eclipse.sirius.components.palette.dto.Palette;
+import org.eclipse.sirius.components.palette.dto.ToolSection;
 import org.eclipse.sirius.components.core.api.IURLParser;
 import org.eclipse.sirius.components.diagrams.description.DiagramDescription;
 import org.eclipse.sirius.components.interpreter.AQLInterpreter;
